Add tests for different elastic versions
